--- src/openjarvis/memory/chunking.py
@@ -1,8 +1,0 @@
-"""Backward-compatibility shim — canonical location is openjarvis.tools.storage.chunking."""
-from openjarvis.tools.storage.chunking import (  # noqa: F401
-    Chunk,
-    ChunkConfig,
-    chunk_text,
-)
-
-__all__ = ["Chunk", "ChunkConfig", "chunk_text"]
