--- src/openjarvis/memory/hybrid.py
@@ -1,7 +1,0 @@
-"""Backward-compatibility shim — canonical location is openjarvis.tools.storage.hybrid."""
-from openjarvis.tools.storage.hybrid import (  # noqa: F401
-    HybridMemory,
-    reciprocal_rank_fusion,
-)
-
-__all__ = ["HybridMemory", "reciprocal_rank_fusion"]
