--- src/openjarvis/memory/sqlite.py
@@ -1,4 +1,0 @@
-"""Backward-compatibility shim — canonical location is openjarvis.tools.storage.sqlite."""
-from openjarvis.tools.storage.sqlite import SQLiteMemory  # noqa: F401
-
-__all__ = ["SQLiteMemory"]
