+++ tests/cli/test_memory_cmd.py
@@ -9,7 +9,7 @@ from click.testing import CliRunner
 
 from openjarvis.cli import cli
 from openjarvis.core.registry import MemoryRegistry
-from openjarvis.memory.sqlite import SQLiteMemory
+from openjarvis.tools.storage.sqlite import SQLiteMemory
 
 
 def _register_sqlite():
