+++ tests/memory/test_bm25.py
@@ -8,7 +8,7 @@ rank_bm25 = pytest.importorskip("rank_bm25")
 
 from openjarvis.core.events import EventBus, EventType  # noqa: E402
 from openjarvis.core.registry import MemoryRegistry  # noqa: E402
-from openjarvis.memory.bm25 import BM25Memory  # noqa: E402
+from openjarvis.tools.storage.bm25 import BM25Memory  # noqa: E402
 
 
 def _make_backend() -> BM25Memory:
@@ -124,7 +124,7 @@ def test_clear():
 def test_event_bus_store():
     bus = EventBus(record_history=True)
     backend = _make_backend()
-    import openjarvis.memory.bm25 as mod
+    import openjarvis.tools.storage.bm25 as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
@@ -145,7 +145,7 @@ def test_event_bus_retrieve():
     bus = EventBus(record_history=True)
     backend = _make_backend()
     backend.store("searchable content for events")
-    import openjarvis.memory.bm25 as mod
+    import openjarvis.tools.storage.bm25 as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
