+++ tests/memory/test_colbert.py
@@ -10,7 +10,7 @@ import torch  # noqa: E402
 
 from openjarvis.core.events import EventBus, EventType  # noqa: E402
 from openjarvis.core.registry import MemoryRegistry  # noqa: E402
-from openjarvis.memory.colbert_backend import ColBERTMemory  # noqa: E402
+from openjarvis.tools.storage.colbert_backend import ColBERTMemory  # noqa: E402
 
 
 def _make_backend() -> ColBERTMemory:
@@ -128,7 +128,7 @@ def test_clear():
 def test_event_bus_store():
     bus = EventBus(record_history=True)
     backend = _make_backend()
-    import openjarvis.memory.colbert_backend as mod
+    import openjarvis.tools.storage.colbert_backend as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
@@ -149,7 +149,7 @@ def test_event_bus_retrieve():
     bus = EventBus(record_history=True)
     backend = _make_backend()
     backend.store("searchable content for events")
-    import openjarvis.memory.colbert_backend as mod
+    import openjarvis.tools.storage.colbert_backend as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
