+++ tests/memory/test_faiss.py
@@ -13,9 +13,9 @@ from openjarvis.core.events import (  # noqa: E402
     EventType,
 )
 from openjarvis.core.registry import MemoryRegistry  # noqa: E402
-from openjarvis.memory._stubs import RetrievalResult  # noqa: E402
-from openjarvis.memory.embeddings import Embedder  # noqa: E402
-from openjarvis.memory.faiss_backend import (  # noqa: E402
+from openjarvis.tools.storage._stubs import RetrievalResult  # noqa: E402
+from openjarvis.tools.storage.embeddings import Embedder  # noqa: E402
+from openjarvis.tools.storage.faiss_backend import (  # noqa: E402
     FAISSMemory,
 )
 
@@ -168,7 +168,7 @@ def test_event_bus_store():
     bus = EventBus(record_history=True)
     backend = _make_backend()
 
-    import openjarvis.memory.faiss_backend as mod
+    import openjarvis.tools.storage.faiss_backend as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
@@ -192,7 +192,7 @@ def test_event_bus_retrieve():
     backend = _make_backend()
     backend.store("searchable content for events")
 
-    import openjarvis.memory.faiss_backend as mod
+    import openjarvis.tools.storage.faiss_backend as mod
 
     original = mod.get_event_bus
     mod.get_event_bus = lambda: bus
