+++ tests/memory/test_ingest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from openjarvis.memory.ingest import (
+from openjarvis.tools.storage.ingest import (
     detect_file_type,
     ingest_path,
     read_document,
